feat (JwtUtil, controller, service) : 필터 없이 유저랑 포스트 매핑

+++ src/main/java/com/sparta_a5/ootd/post/service/PostService.java
@@ -4,6 +4,7 @@ import com.sparta_a5.ootd.post.dto.PostRequestDto;
 import com.sparta_a5.ootd.post.dto.PostResponseDto;
 import com.sparta_a5.ootd.post.entity.Post;
 import com.sparta_a5.ootd.post.repository.PostRepository;
+import com.sparta_a5.ootd.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,9 +17,11 @@ public class PostService {
 
     private final PostRepository postRepository;
 
+
     @Transactional
-    public PostResponseDto createPost(PostRequestDto postRequestDto) {
-        Post post = postRepository.save(new Post(postRequestDto));
+    public PostResponseDto createPost(PostRequestDto postRequestDto, User user) {
+
+        Post post = postRepository.save(new Post(postRequestDto,user));
         return new PostResponseDto(post);
     }
 
@@ -39,19 +42,27 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponseDto updatePost(Long postId, PostRequestDto postRequestDto) {
+    public PostResponseDto updatePost(Long postId, PostRequestDto postRequestDto, User user) {
          Post post = postRepository.findById(postId).orElseThrow(
                  ()-> new IllegalArgumentException("존재하지 않는 글입니다.")
          );
 
-         post.update(postRequestDto);
-         return new PostResponseDto(post);
+         if(post.getUser() == user){
+             post.update(postRequestDto);
+             return new PostResponseDto(post);
+         } else {
+             throw new IllegalArgumentException("권한이 없습니다.");
+         }
     }
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId,User user) {
         Post post = postRepository.findById(postId).orElseThrow(
                 ()-> new IllegalArgumentException("존재하지 않는 글입니다.")
         );
-        postRepository.delete(post);
+        if(post.getUser() == user){
+            postRepository.delete(post);
+        } else {
+            throw new IllegalArgumentException("권한이 없습니다.");
+        }
     }
 }
